Remove duplicate version of format_uncertainty
